# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Legal Copilot API
  version: 1.0.0
  description: |-
    AI-first practice management API for UK law firms.

    ## Authentication
    All endpoints require authentication via Bearer token.

    ## Rate Limiting
    API requests are rate limited to 100 requests per minute.

    ## Pagination
    List endpoints support pagination via `page` and `limit` query parameters.
  contact:
    name: Legal Copilot Support
servers:
  - url: http://localhost:3000
    description: Development server
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: Validation failed
        code:
          type: string
          description: Machine-readable error code
          example: VALIDATION_ERROR
        details:
          type: object
          additionalProperties:
            nullable: true
          description: Additional error details
      required:
        - error
...
    Matter:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier in UUID v4 format
          example: 123e4567-e89b-12d3-a456-426614174000
        reference:
          type: string
          example: SMI-2024-0042
        title:
          type: string
          example: Smith v Jones - Property Dispute
        description:
          type: string
          nullable: true
        clientId:
          type: string
          format: uuid
          description: Unique identifier in UUID v4 format
          example: 123e4567-e89b-12d3-a456-426614174000
        feeEarnerId:
          type: string
          nullable: true
          format: uuid
          description: Unique identifier in UUID v4 format
          example: 123e4567-e89b-12d3-a456-426614174000
        supervisorId:
          type: string
          nullable: true
          format: uuid
          description: Unique identifier in UUID v4 format
          example: 123e4567-e89b-12d3-a456-426614174000
        status:
          type: string
          enum: &a6
            - lead
            - active
            - on_hold
            - closed
            - archived
          description: Current status in matter lifecycle
          example: active
        practiceArea:
          type: string
          enum: &a4
            - conveyancing
            - litigation
            - family
            - probate
            - employment
            - immigration
            - personal_injury
            - commercial
            - criminal
            - ip
            - insolvency
            - other
          description: Legal practice area
          example: conveyancing
        billingType:
          type: string
          enum: &a5
            - hourly
            - fixed_fee
            - conditional
            - legal_aid
            - pro_bono
          description: Billing arrangement type
          example: hourly
        hourlyRate:
          type: string
          nullable: true
          pattern: ^\d+\.\d{2}$
          description: Monetary amount with 2 decimal places
          example: "1250.00"
        fixedFee:
          type: string
          nullable: true
          pattern: ^\d+\.\d{2}$
          description: Monetary amount with 2 decimal places
          example: "1250.00"
        estimatedValue:
          type: string
          nullable: true
          pattern: ^\d+\.\d{2}$
          description: Monetary amount with 2 decimal places
          example: "1250.00"
        openedAt:
          type: string
          nullable: true
          format: date-time
          description: ISO 8601 date-time string
          example: 2024-12-17T10:30:00Z
        closedAt:
          type: string
          nullable: true
          format: date-time
          description: ISO 8601 date-time string
          example: 2024-12-17T10:30:00Z
        keyDeadline:
          type: string
          nullable: true
          format: date-time
          description: ISO 8601 date-time string
          example: 2024-12-17T10:30:00Z
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
          description: ISO 8601 date-time string
          example: 2024-12-17T10:30:00Z
        updatedAt:
          type: string
          format: date-time
          description: ISO 8601 date-time string
          example: 2024-12-17T10:30:00Z
      required:
        - id
        - reference
        - title
        - description
        - clientId
        - feeEarnerId
        - supervisorId
        - status
        - practiceArea
        - billingType
        - hourlyRate
        - fixedFee
        - estimatedValue
        - openedAt
        - closedAt
        - keyDeadline
        - notes
        - createdAt
        - updatedAt
    CreateMatterRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
        clientId:
          type: string
          format: uuid
          description: Unique identifier in UUID v4 format
          example: 123e4567-e89b-12d3-a456-426614174000
        feeEarnerId:
          type: string
          format: uuid
          description: Unique identifier in UUID v4 format
          example: 123e4567-e89b-12d3-a456-426614174000
        supervisorId:
          type: string
          format: uuid
          description: Unique identifier in UUID v4 format
          example: 123e4567-e89b-12d3-a456-426614174000
        practiceArea:
          type: string
          enum: *a4
          description: Legal practice area
          example: conveyancing
        billingType:
          type: string
          enum: *a5
          default: hourly
          description: Billing arrangement type
          example: hourly
        hourlyRate:
          type: string
        fixedFee:
          type: string
        estimatedValue:
          type: string
        keyDeadline:
          type: string
          format: date-time
        notes:
          type: string
      required:
        - title
        - clientId
        - practiceArea
...
    MatterListResponse:
      type: object
      properties:
        matters:
          type: array
          items:
            $ref: "#/components/schemas/Matter"
        pagination:
          type: object
          properties:
            page:
              type: number
            limit:
              type: number
            total:
              type: number
            totalPages:
              type: number
          required:
            - page
            - limit
            - total
            - totalPages
      required:
        - matters
        - pagination
...
    DocumentType:
      type: string
      enum:
        - letter_in
        - letter_out
        - email_in
        - email_out
        - contract
        - court_form
        - evidence
        - note
        - id_document
        - financial
        - other
    DocumentStatus:
      type: string
      enum:
        - draft
        - pending_review
        - approved
        - sent
        - archived
    Document:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier in UUID v4 format
          example: 123e4567-e89b-12d3-a456-426614174000
        matterId:
          type: string
          format: uuid
          description: Unique identifier in UUID v4 format
          example: 123e4567-e89b-12d3-a456-426614174000
        title:
          type: string
        type:
          $ref: "#/components/schemas/DocumentType"
        status:
          $ref: "#/components/schemas/DocumentStatus"
        uploadId:
          type: string
          nullable: true
          format: uuid
          description: Unique identifier in UUID v4 format
          example: 123e4567-e89b-12d3-a456-426614174000
        filename:
          type: string
          nullable: true
        mimeType:
          type: string
          nullable: true
        fileSize:
          type: integer
          nullable: true
        createdBy:
          type: string
          nullable: true
          format: uuid
          description: Unique identifier in UUID v4 format
          example: 123e4567-e89b-12d3-a456-426614174000
        documentDate:
          type: string
          nullable: true
          format: date-time
          description: ISO 8601 date-time string
          example: 2024-12-17T10:30:00Z
        recipient:
          type: string
          nullable: true
        sender:
          type: string
          nullable: true
        aiSummary:
          type: string
          nullable: true
        extractedText:
          type: string
          nullable: true
        chunkedAt:
          type: string
          nullable: true
          format: date-time
          description: ISO 8601 date-time string
          example: 2024-12-17T10:30:00Z
        chunkCount:
          type: integer
          nullable: true
        metadata:
          type: object
          nullable: true
          additionalProperties:
            nullable: true
        version:
          type: integer
        parentId:
          type: string
          nullable: true
          format: uuid
          description: Unique identifier in UUID v4 format
          example: 123e4567-e89b-12d3-a456-426614174000
        createdAt:
          type: string
          format: date-time
          description: ISO 8601 date-time string
          example: 2024-12-17T10:30:00Z
        updatedAt:
          type: string
          format: date-time
          description: ISO 8601 date-time string
          example: 2024-12-17T10:30:00Z
      required:
        - id
        - matterId
        - title
        - type
        - status
        - uploadId
        - filename
        - mimeType
        - fileSize
        - createdBy
        - documentDate
        - recipient
        - sender
        - aiSummary
        - extractedText
        - chunkedAt
        - chunkCount
        - metadata
        - version
        - parentId
        - createdAt
        - updatedAt
    DocumentListResponse:
      type: object
      properties:
        documents:
          type: array
          items:
            $ref: "#/components/schemas/Document"
      required:
        - documents
    CreateDocumentRequest:
      type: object
      properties:
        matterId:
          type: string
          format: uuid
          description: Unique identifier in UUID v4 format
          example: 123e4567-e89b-12d3-a456-426614174000
        title:
          type: string
          minLength: 1
          maxLength: 200
        type:
          $ref: "#/components/schemas/DocumentType"
        uploadId:
          type: string
          format: uuid
          description: Unique identifier in UUID v4 format
          example: 123e4567-e89b-12d3-a456-426614174000
        filename:
          type: string
        mimeType:
          type: string
        fileSize:
          type: integer
        documentDate:
          type: string
          format: date-time
          description: ISO 8601 date-time string
          example: 2024-12-17T10:30:00Z
        recipient:
          type: string
        sender:
          type: string
        extractedText:
          type: string
        metadata:
          type: object
          additionalProperties:
            nullable: true
      required:
        - matterId
        - title
        - type
...
    TimelineEventType:
      type: string
      enum:
        - matter_created
        - matter_updated
        - matter_archived
        - document_uploaded
        - document_chunked
        - task_created
        - task_completed
        - approval_decided
        - note_added
    TimelineActorType:
      type: string
      enum:
        - user
        - system
        - ai
    TimelineEvent:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier in UUID v4 format
          example: 123e4567-e89b-12d3-a456-426614174000
        matterId:
          type: string
          format: uuid
          description: Unique identifier in UUID v4 format
          example: 123e4567-e89b-12d3-a456-426614174000
        type:
          $ref: "#/components/schemas/TimelineEventType"
        title:
          type: string
        description:
          type: string
          nullable: true
        actorType:
          $ref: "#/components/schemas/TimelineActorType"
        actorId:
          type: string
          nullable: true
          format: uuid
          description: Unique identifier in UUID v4 format
          example: 123e4567-e89b-12d3-a456-426614174000
        entityType:
          type: string
          nullable: true
        entityId:
          type: string
          nullable: true
          format: uuid
          description: Unique identifier in UUID v4 format
          example: 123e4567-e89b-12d3-a456-426614174000
        metadata:
          type: object
          nullable: true
          additionalProperties:
            nullable: true
        occurredAt:
          type: string
          format: date-time
          description: ISO 8601 date-time string
          example: 2024-12-17T10:30:00Z
        createdAt:
          type: string
          format: date-time
          description: ISO 8601 date-time string
          example: 2024-12-17T10:30:00Z
      required:
        - id
        - matterId
        - type
        - title
        - description
        - actorType
        - actorId
        - entityType
        - entityId
        - metadata
        - occurredAt
        - createdAt
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
      required:
        - page
        - limit
        - total
        - totalPages
        - hasNext
        - hasPrev
    TimelineListResponse:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/TimelineEvent"
        pagination:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - events
        - pagination
...
    CreateTimelineEventRequest:
      type: object
      properties:
        type:
          allOf:
            - $ref: "#/components/schemas/TimelineEventType"
            - default: note_added
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 10000
        occurredAt:
          type: string
          format: date-time
          description: ISO 8601 date-time string
          example: 2024-12-17T10:30:00Z
        metadata:
          type: object
          additionalProperties:
            nullable: true
      required:
        - title
...
paths:
...
  /api/matters:
    get:
      summary: List matters
      description: Retrieve a paginated list of matters
      tags:
        - Matters
      parameters:
        - schema:
            type: integer
            minimum: 1
            default: 1
          required: false
          name: page
          in: query
        - schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          required: false
          name: limit
          in: query
        - schema:
            type: string
            enum: *a6
            description: Current status in matter lifecycle
            example: active
          required: false
          name: status
          in: query
        - schema:
            type: string
            enum: *a4
            description: Legal practice area
            example: conveyancing
          required: false
          name: practiceArea
          in: query
        - schema:
            type: string
            format: uuid
            description: Unique identifier in UUID v4 format
            example: 123e4567-e89b-12d3-a456-426614174000
          required: false
          name: clientId
          in: query
        - schema:
            type: string
            format: uuid
            description: Unique identifier in UUID v4 format
            example: 123e4567-e89b-12d3-a456-426614174000
          required: false
          name: feeEarnerId
          in: query
        - schema:
            type: string
            description: Search by reference or title
          required: false
          name: search
          in: query
      responses:
        "200":
          description: List of matters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatterListResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Create matter
      description: Create a new matter/case
      tags:
        - Matters
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMatterRequest"
      responses:
        "201":
          description: Matter created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Matter"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/documents:
    get:
      summary: List documents
      description: Retrieve documents, optionally filtered by matterId
      tags:
        - Documents
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentListResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Create document
      description: Create a new document record for a matter
      tags:
        - Documents
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentRequest"
      responses:
        "201":
          description: Document created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/matters/{id}/timeline:
    get:
      summary: List timeline events
      description: Retrieve timeline events for a matter
      tags:
        - Timeline
      parameters:
        - schema:
            type: string
            format: uuid
            description: Unique identifier in UUID v4 format
            example: 123e4567-e89b-12d3-a456-426614174000
          required: true
          name: id
          in: path
        - schema:
            type: integer
            minimum: 1
            default: 1
            description: Page number (1-indexed)
            example: 1
          required: false
          name: page
          in: query
        - schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
            description: Items per page (max 100)
            example: 20
          required: false
          name: limit
          in: query
        - schema:
            $ref: "#/components/schemas/TimelineEventType"
          required: false
          name: type
          in: query
        - schema:
            type: string
          required: false
          name: entityType
          in: query
        - schema:
            type: string
            format: uuid
            description: Unique identifier in UUID v4 format
            example: 123e4567-e89b-12d3-a456-426614174000
          required: false
          name: entityId
          in: query
        - schema:
            type: string
            format: date-time
            description: ISO 8601 date-time string
            example: 2024-12-17T10:30:00Z
          required: false
          name: from
          in: query
        - schema:
            type: string
            format: date-time
            description: ISO 8601 date-time string
            example: 2024-12-17T10:30:00Z
          required: false
          name: to
          in: query
      responses:
        "200":
          description: Timeline events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimelineListResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Create timeline event
      description: Create a manual timeline event for a matter
      tags:
        - Timeline
      parameters:
        - schema:
            type: string
            format: uuid
            description: Unique identifier in UUID v4 format
            example: 123e4567-e89b-12d3-a456-426614174000
          required: true
          name: id
          in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTimelineEventRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimelineEvent"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...